feat(api): This commit includes a new payment microservice feature, and this is its initial commit.
The payment microservice allows for debiting users from a Supabase user database table, and correctly tracks transactions with a transaction table also required to be present within the Supabase directory. The endpoints directly communicate with a RPC function which also is required to be present.

Improvements: adding a robust test-suite is a valid improvement for the furture as it is yet to be created.